fix(0231): 2 的幂，循环方案，比较 Double 和 Int 值时，将 Double 转为 Int 比较，导致计算结果错误。

+++ src/main/kotlin/com/github/xumeng/kotlin/problems/easy/problem0231/Solution0231S1.kt
@@ -9,8 +9,8 @@ class Solution0231S1 {
             return false
         }
         for (i in 0..<32) {
-            val power = 2.0.pow(i).toInt()
-            if (power == n) {
+            val power = 2.0.pow(i)
+            if (power == n.toDouble()) {
                 return true
             }
             if (power > n) {
